Add support to OSX project

+++ scripts/remove-key.sh
@@ -4,13 +4,16 @@ if [[ "$TRAVIS_PULL_REQUEST" != "false" ]]; then
     exit 0
 fi
 
-security delete-keychain ios-build.keychain
-rm -f "~/Library/MobileDevice/Provisioning Profiles/$PROFILE_NAME.mobileprovision"
-if [[ -n "$APP_EXTENSION_PROFILE_NAME" ]]; then
-	rm -f "~/Library/MobileDevice/Provisioning Profiles/$APP_EXTENSION_PROFILE_NAME.mobileprovision"
-  rm "profile/$APP_EXTENSION_PROFILE_NAME.mobileprovision"
-fi
+security delete-keychain ${PROJECT_TYPE}-build.keychain
 
-rm "profile/$PROFILE_NAME.mobileprovision"
-rm "certs/dist.cer"
-rm "certs/dist.p12"
+if [[ "$PROJECT_TYPE" == "ios" ]]; then
+  rm -f "~/Library/MobileDevice/Provisioning Profiles/$PROFILE_NAME.mobileprovision"
+  if [[ -n "$APP_EXTENSION_PROFILE_NAME" ]]; then
+    rm -f "~/Library/MobileDevice/Provisioning Profiles/$APP_EXTENSION_PROFILE_NAME.mobileprovision"
+    rm "profile/$APP_EXTENSION_PROFILE_NAME.mobileprovision"
+  fi
+
+  rm "profile/$PROFILE_NAME.mobileprovision"
+  rm "certs/dist.cer"
+  rm "certs/dist.p12"
+fi
